// src/types/Monitoring.types.d.ts
export interface betweenDatesType {
  dates: string[];
  timeStamps: number[];
  lastTimestamp: number;
  firstTimestamp: number;
}

export interface deviceHistoryType {
  id: number;
  time: string;
  value: string;
  status: string;
  key: string;
}

export interface deviceHistoriesType {
  name: string;
  telemetryDataPoints: {
    date: string;
    noise: string;
    status: string;
    temperature: string;
    vibration: string;
    id: string;
  }[];
}
// ...
export interface recordsRangeType {
  start: number;
  end: number;
}